Add Ben's Data
Ben from https://nbviewer.org/url/argmin.net/code/revisiting-bd-mask-rct.ipynb
did some analysis on endlineBlook_Data.dta which was generated by the
original paper author's code. However that relied on stata to generate
and we don't have that available.  So Ben sent me the file and I also
converted it to csv.  That's included here for everyone as an open format.